requirements.txt
google-cloud-storage==3.6.0
pillow==12.0.0
